+++ trace/table.go
@@ -96,7 +96,7 @@ type (
 	}
 	tableSessionInfo interface {
 		ID() string
-		NodeID() int64
+		NodeID() uint32
 		Status() string
 		LastUsage() time.Time
 	}
